fixed seeders and fixed paymentcontroller

+++ database/migrations/2021_12_02_151316_create_sponsors_table.php
@@ -17,7 +17,7 @@ class CreateSponsorsTable extends Migration
             
             $table->id();
             $table->string('name', 100)->unique();
-            $table->tinyInteger('duration');
+            $table->smallInteger('duration');
             $table->float('price', 10, 2);
             $table->timestamps();
         });
